Use collectTransitive instead of collect in PyCommon
The instance of an anonymous subclass of LocalMetadataCollector used in this call currently doesn't collect anything. Better to use the shorter call that doesn't handle metadata collection.

(This doesn't fix the issue that collectTransitive is a confusing name. It only calls collect with reasonable default values for absent parameters, and probably should just have the same name as other overloads of InstrumentedFileCollector.collect.)

RELNOTES: None.
PiperOrigin-RevId: 352813404