Automated rollback of commit a8f6152e9f62cd65aded6df3723d0bbb44d4603f.
*** Reason for rollback ***

Causes blaze crash, discovered in TGP for 2021.01.20 nightly: b/178126050

*** Original change description ***

Refactors CompilationSupport for objc to use existing API

This CL makes CompilationSupport use CcLinkingHelper instead of CppLinkActionBuilder. The former is the Java class used by the existing Starlark linking API.

This is in preparation for a future (unknown when) re-write of objc rules to
Starlark.

PiperOrigin-RevId: 353142462