Remote: Use parameters instead of thread-local storage to provide tracing metadata.
Remote Execution API defines RequestMetadata which is an optional message attached to any RPC request to tell the server about an external context of the request. Existing code uses Context object from gRPC which is essentially a thread-local storage to provide metadata:
    1. We always attach the Context even the underlying implementation isn't based on gRPC.
    2. It is error prone in a multi-threaded context.

This is the first step of a series changes where we introduce RemoteActionExecutionContext and update RemoteCacheClient#downloadActionResult to use it.

These is a regression that networkTime of SpawnMetrics will not be correct since the NetworkTimeInterceptor is no longer installed at RemoteModule and other methods haven't get updated to use the RemoteActionExecutionContext. The networkTime will back to normal once we finish the cleanup.

PiperOrigin-RevId: 353819916