Last bug fix for the starlark build settings x repositories clean up.
This brings us home to the principle of "we always deal with canonicalized settings on the back end". Before this CL, this situation described in https://github.com/bazelbuild/bazel/issues/10499 was still an issue.

Some background:

By design, the label-object map of starlark options in BuildOptions does not store starlark flags that are set to their default value (set either by never having been set, or explicitly set to the default value). This means before we do starlark transitions, we fill out the configuration with all the relevant default values so they're there if we need to read them. Before this CL, those editions to the map were being added in their given form, not there canonical form. This obviously caused issues when our transition logic tried to find these setting using their canonical form. This CL fixes that. The affected logic is ConfigurationResolver#addDefaultStarlarkOptions.

Fixes #11128

PiperOrigin-RevId: 353923997