Clean up the confusing distinction between "BaseRule" and "RuleBase".
Summary:
 - RootRule -> <removed>. The most basic rule definition, just adds the
     "name" attribute. Moved directly to RuleClass.Builder.
 - BaseRule -> NativeBuildRule = all non-Starlark BUILD rules. Includes rules
     that create actions (like compilation) and all other purposes (like aliases,
     filegroups, toolchain definitions that only provide metadata to other rules)
 - RuleBase -> NativeActionCreatingRule = all non-Starlark BUILD rules that
     create actions

Starlark rule classes are defined in StarlarkRuleClassFunctions. These remain unchanged.

This is a best-effort attempt. There seem to be some exceptions to every case,
but this covers the overarching themes as clearly as I can see them (code search
for "RuleBase.class)" and "BaseRule.class)").

Even if not 100% accurate, this communicates a standard to aspire to and make
adjustments around. This alone is a good step up from the current wording.

** See BaseRuleClasses.java for the main change. **

PiperOrigin-RevId: 353944214