Simplify Skyframe "inlining" code path in BzlLoadFunction
(Skyframe "inlining" is when the logic of one Skyfunction occurs in the same thread as another, without a real Skyframe call.)

The BzlLoadFunction#computeInline code path is a rather complicated dance. It entails wrapping and unwrapping the Skyframe environment with instrumentation that logs dependency metadata to nodes in a cache DAG. The timing of these dovetailed recursive steps is somewhat error-prone.

For instance, I recently introduced an error in StarlarkBuiltinsFunction in https://github.com/bazelbuild/bazel/commit/60a4f9e9c1c3898ca846177f70c3045b0c20824d, by requesting a Skyframe value (the StarlarkSemantics) in the no-man's land between unwrapping and re-wrapping the instrumented environment. This error happens to be benign because it's rare (impossible?) for the StarlarkSemantics to be unavailable, so its dependency metadata isn't very important. But this illustrates the potential for danger as the code evolves in the future.

This CL avoids this problem by 1) moving management of this wrapping and unwrapping into the opaque InliningState object, and 2) establishing the invariant that, during inlining, the environment is always wrapped. In the very initial part of the computation, where there is no need to record any dep metadata yet, we still use a trivial wrapped environment for the sake of uniformity.

The `env` param to computeInline() and its helpers is now redundant with the `inliningState` param, and is removed.

A follow-up will move some of the CachedBzlLoadData management into InliningState as well, to simplify computeInlineForCacheMiss().

Cleanup on way to #11437.

PiperOrigin-RevId: 354954719