Fix a small space leak in `TestResultAggregator`.
This is measurable with high values of `shard_count` and/or `runs_per_test`. `TestResultAggregator` unnecessarily retains every `TestResult` in a map even though they are never needed again. For most cases, this is insignificant because all of the test's action(s) will complete within a short time frame. However, when there are many runs, the lifetime and number of results retained by `TestResultAggregator` increases.

The `remainingRuns` set is also unnecessary. All we need to know is:
* Whether *all* test runs have reported a result. For this, we use an `int` that starts at the number of runs and counts down.
* Whether we have already posted a test summary, e.g. because the test failed to build or was skipped. For this, a `boolean` suffices.

PiperOrigin-RevId: 355751321