Quarantine post "--" flag alias unwrapping to canonicalize-command.
Blaze instantiates an options parser (OP-blaze) and the CanonicalizeCommand instantiates an options parser (OP-canonicalize). Add the aliases from OP-blaze parsing to OP-blaze's OptionsParsingResult. These can then be used to seed the creation of OP-canonicalize so that OP-canonicalize can use them.

Modify the OptionsParserImpl to be able to take in an alias map (default empty).

PiperOrigin-RevId: 355901224