Fix Bazel #10214: JDK 13 introduced a source compatibility issue.
Quote from the Java release notes:

The addition of newFileSystem(Path, Map<String, ?>) creates a source (but not binary) compatibility issue for code that has been using the existing 2-arg newFileSystem(Path, ClassLoader) and specifying the class loader as null. [...] To avoid the ambiguous reference, this code needs to be modified to cast the second parameter to java.lang.ClassLoader.

RELNOTES:
PiperOrigin-RevId: 356301318