docs: update java.bzl best practices to rm dead URL
The rules for dealing with external java dependencies offer best practices in a
comment that includes a note about using `maven.ibiblio.org` as a third
`jar_urls` entry:

https://github.com/bazelbuild/bazel/blob/81a5a12b3566e0b341c3cfbc271a60e568e1ed41/tools/build_defs/repo/java.bzl#L47

At the time of writing this domain consistently fails to resolve w/ status
`NXDOMAIN`:

```bash
 $> dig +noall +comments +answer  @8.8.8.8 maven.ibiblio.org
;; Got answer:
;; ->>HEADER<<- opcode: QUERY, status: NXDOMAIN, id: 19083
;; flags: qr rd ra; QUERY: 1, ANSWER: 0, AUTHORITY: 1, ADDITIONAL: 1

;; OPT PSEUDOSECTION:
; EDNS: version: 0, flags:; udp: 512
```

This commit removes the `maven.ibiblio.org` URLs throughout. I'm not familiar
enough with the Java packaging ecosystem in 2021 to suggest an alternative but
leaving it as is produces unnecessary warning logs when the dead mirror is
selected from the `jar_urls` for a dep.