Build Bazel explicitly with Xcode 11.7.
This is currently a no-op, because the default version of Xcode on Bazel CI is 11.7 anyway.

We'll upgrade to Xcode 12.4 shortly and also make that the default version on Bazel CI, but having a commit with an explicit Xcode version for 11.7 will make it easier in case we have to cherry-pick this into 3.7.x or 4.x.

PiperOrigin-RevId: 356965864