Stop generating compile info in ObjcProvider
Step 3 of 3-step plan:
1. Migrate the definitions.
2. Migrate the uses.
3. Delete old ObjcProvider definitions.

See https://github.com/bazelbuild/bazel/issues/10674.

Now that compilation info has been migrated from ObjcProvider to
CcCompilationContext, get rid of some logic used for the migration.

Before the migration, ObjcCommon.build() produced an ObjcProvider.

For the migration, I added a CcCompilationContext inside the
ObjcProvider.  This caused two issues:

1. the CcCompilationContext may not exist during ObjcCommon.build(),
   so it would not be possible to fully construct ObjcProvider then.

2. Depending on the use case, CcCompilationContext may be produced in
   two different ways.

So I changed ObjcCommon to produce a Builder for ObjcProvider instead,
with its CcCompilationContext to be set later when it is available.

Now that the migration is complete, we no longer need the
CcCompilationContext in ObjcProvider, so delete all this logic, and
revert to having ObjcCommon generate an ObjcProvider with an empty
CcCompilationContext.  I will delete ObjcProvider's
CcCompilationContext in a subsequent CL.

I also had to update the tests so that they use CcInfo instead of
ObjcProvider to find objc rules' compile info.

PiperOrigin-RevId: 357642600