Switch from `getSegments()` to `containsUpLevelReferences()` in `ArtifactRoot#asDerivedRoot`.
`getSegments()` iterates over the path string twice and also creates garbage by taking substrings and putting them in a list. `containsUpLevelReferences()` takes constant time, as it does not need to look past the 3rd character of the already-normalized path string.

Additionally, clean up `PathFragment#containsUpLevelReferences` so that it does not perform string concatenation.

PiperOrigin-RevId: 357826526