Fix link text
I realize the link starts with be/ but from the user experience POV, that link doesn't refer into the build encyclopedia, at least not according to the TOC in the left-hand column.  If you don't want to accept this change, perhaps the TOC should be changed to reflect the file structure of the site?

Closes #13027.

PiperOrigin-RevId: 357834652