Optimize `InMemoryFileSystem#pathWalkErrno`.
The main culprit is using `getParentDirectory` and `baseNameOrWindowsDrive` to obtain a stack of path segments. This creates a huge amount of garbage.

Take for example `path = /a/b/c/d/e`. The current approach does:
- `push("e"), path = "/a/b/c/d"`
- `push("d"), path = "/a/b/c"`
- `push("c"), path = "/a/b"`
- `push("b"), path = "/a"`
- `push("a"), path = "/"`

In addition to creating the path segments that are actually needed, we create garbage that is quadratic in the length of the original path. Instead, introduce a `PathSegmentIterator` that lazily iterates over the path string, yielding segments when it encounters a separator character. I suspect that other places can benefit from this approach, so I filed b/180422294.

PiperOrigin-RevId: 358072243