Allow using `PathSegmentIterator` to iterate over `PathFragment` segments.
This is an initial pass to find callers of `PathFragment#get(First)Segment(s)` that can clearly and easily benefit from doing something smarter. There are plenty more places that can be improved, but will take more work.

Since there are no more callers of the method that returned an array of segments, have `getSegments()` populate an `ImmutableList.Builder` directly. To make sure it is a strict performance improvement, continue to count the segments first to accurately pre-size the builder.

PiperOrigin-RevId: 358204210