Fix getting started/doc nav split for published pages
Restore the documentation sidebar for published doc versions, since they are all categorized under documentation, not getting started. This fixes the problem where the current 4.0 version bazel-overview.html loads the documentation sidebar, although categorized in getting-started on master.

This also means that the getting-started/documentation navbar split is only visible on master, not 4.0 and before.

Closes #13063.

PiperOrigin-RevId: 358320934