Rename `PathFragment#getSegments` to `splitToListOfSegments` and perform another pass to migrate unnecessary callers.
The renaming makes it clearer that it is not a simple getter method, but rather one that requires string operations that can be costly in aggregate.

There is an overload of `ArtifactRoot#asDerivedRoot` where callers seem to unnecessarily split up a path fragment only for it to reconstruct one and call the other overload. Those calls are fixed to use the more efficient overload.

Switched some test calls to `segments()` where the assertion is equivalent. Hopefully this will encourage that method to be preferred.

PiperOrigin-RevId: 359761371