Clean up some things in `InMemoryFilesystem`.
Permit calling `getxattr` on a symlink. The `FileSystem` documentation incorrectly claims that it follows symlinks, but there is actually a boolean parameter that determines this. Change it from throwing `IllegalStateException` from the `instanceof` check to just returning `null`.

Reduce unnecessary calls to `Path#isReadable`. There were several spots calling this and duplicating a call to `inodestat` even though `InMemoryContentInfo#isReadable` can be checked directly. This is more evidence that `Path` should not be passed to `FileSystem` methods.

PiperOrigin-RevId: 359834173