Do not pass an ArtifactExpander to CompletionContext.
Instead, pass in the required maps because the behavior of CompletionContext is simple enough that it makes more sense to do that instead.

The long-term goal is to turn the completion events into data objects. Unfortunately, that probably won't happen for a while since pathResolver does to useful work and it's hard to plumb that in in any other way.

RELNOTES: None.
PiperOrigin-RevId: 360205871