Disable layering check and parse headers for J2Objc transpiled code
In general, there is less value in doing these checks for generated
code -- e.g. we don't do it for proto headers.

Parse headers is currently off for J2Objc anyways, but it would
otherwise be enabled when we migrate header tokens to
CcCompilationContext.

Layering check is not currently available, but when it is, we would
like it to continue disabling it for transpiled code.

PiperOrigin-RevId: 360227588