Batch results from futures of expressions within `query deps(...)` so that `GraphlessBlazeQueryEnvironment#deps` is only called once.
`GraphlessBlazeQueryEnvironment#deps` relies on it being only run once when the expression has been completely evaluated as it does heavy blocking work. `GraphlessBlazeQueryEnvironment#eval` assumes that all expressions only generate a single `QueryTaskFuture` which is not the case for `SetExpression`.

By adding an option to `GraphlessBlazeQueryEnvironment#eval` to decide whether to batch, this fixes the issue by ensuring that the results of futures from expressions within `deps` (when using graphless) are batched before running the callback on the complete result.

PiperOrigin-RevId: 360467695