Reduce amount of garbage produced when collecting directories to create for in sandboxed exec root.
`AbstractContainerizingSandboxedSpawn::createDirectories` constructs a set of
directories to be created in order to perform `mkdir` exactly once for each
directory to be created. In order to create parent directories before children,
we construct a `LinkedHashSet` in particular order. The problem with this
approach is that we create a `Path` object every possible directory on the path
to the sandbox exec root. That could be avoided.

Imagine creating directory `1/2/3/4/`, that would require us to create `1/`,
`1/2/`, `1/2/3/` and `1/2/3/4/`. If we already knew that we are creating
`1/2/3/` for another file/directory, we could avoid creating all `Path` objects
for all of the ancestors. This way, we could reduce the amount of garbage
created in the routine.

Change how we create directories not to require a set ordered in topological
order. Optimize out creating entries for ancestor directories if we are created
for another entry.

PiperOrigin-RevId: 360668532