Add logging for suppressed IOExceptions when getting values via SkyFunctionEnvironment.
This helps to see how such IOExceptions can bubble up to unexpected places. Because I expect such unexpected exceptions to be hit serially in quick succession, I didn't put a rate-limiter in. It would be nice if Flogger had an "atMostNEverySecond()" modifier so that I could rate-limit this but allow a few in quick succession.

PiperOrigin-RevId: 360948592