Test if JavaInfo identical to one returned by Java rules can be constructed.
I implemented the tests for the most basic parameters of java_library, that is srcs, deps, runtime_deps, exports. More complex examples with annotation processing are to follow.

Comparing JavaInfo provider in Java was challenging, because most of the Bazel classes (providers, NestedSets) implement equals based on reference. So in order to compare it you need to destruct the object down to NestedSets and then compare each Artifact based on path. Artifacts equals method actually depends on the producing action (so they would be different), while compareTo==0 only compares path. The whole JavaInfo comparison in Java would result in a lot of boilerplate code.

This is why I chose to use Starlark to convert a JavaInfo object into Dictionary with string values (JSON like - can have other Dicts and Lists). Those are easier to compare and they produce nice assertion messages. Also note that JavaInfo object is completely inspectable in Starlark, so the comparison is also complete.

Those tests already discovered two problem, that will be fixed in following CLs:
- cannot set value of native_headers or manifest
- JavaExportsProvider is not set with exports parameter (but other values in JavaInfo are already set correctly)

PiperOrigin-RevId: 361131344