Emit newError in EvaluationProgressReceiver, clean it up
Implementations of EvaluationProgressReceiver may now be informed about
new error values which nodes evaluate to.

This change also cleans up that interface's documentation, which had
drifted.

This change is separate from a forthcoming change that makes use of this
newly emitted data, because of the low signal-to-noise property of this
change.

The files with significant changes include:
- EvaluationProgressReceiver
- AbstractExceptionalParallelEvaluator

RELNOTES: None.
PiperOrigin-RevId: 361217625