Replace `Path` with `PathFragment` in `FileSystem` interface.
`FileSystem` interface currently takes a `Path` object for all of its
operations. `Path` internally contains a `FileSystem` reference and the path
string itself. Passing `this` instance as part of the parameter makes the
interface more complicated than needed (in fact, we would never use that).
This, for example, results with unnecessary complexity in `DelegateFileSystem`
which recreates the `Path` objects just to call methods on the delegate.

Refactor the `FileSystem` interaface to take a `PathFragment` instead.

PiperOrigin-RevId: 361569044