More properly destroy workers on interrupt.
Interrupts did not wake the workers out of waiting for a response. When running under dynamic execution, that's all we can do until cancellation is implemented. For workers without dynamic execution, the only interrupt is from the build itself getting interrupted, and in that case we don't want to wait - it could take minutes. Instead we then use a busy-wait to notice interrupts and destroy the worker if interrupted. This will be useful even when cancellation is implemented, since not all workers may be able to implement a sensible cancellation.

This also renames mustBeSandboxed on WorkerKey to isSpeculative to separate cause and effect.

RELNOTES: None.
PiperOrigin-RevId: 361802760