Spawn execution log: add wall time duration
The spawn execution log is useful to compare builds, for example to
discover actions that were not cached, and to debug why they were not
cached.

This change adds the wall time of each spawn, which further allows
finding actions that are are faster or slower between two builds.
For example, when comparing local and remote builds, this helps
discover differences in execution time, particularly if some actions
are using persistent workers.

Change-Id: Iacda7cd5b0265ec26d0a8fea82db2690582ed922