Migrate swift module map to its own action
This is in preparation for enabling layering check for `objc_library`.

Layering check requires that a module map be generated.  Currently,
`objc_library` generates a module map to be used by `swift_library`.
The two module maps are incompatible, so we migrate the swift one to
its own action, freeing up the one generated by the cpp to be used for
layering check.

Eventually, the module map needed by `swift_library` will be migrated
to `swift_library`.

PiperOrigin-RevId: 362706906