SpawnAction.Builder: use callable PathFragment
If you pass an Artifact or FilesToRunProvider to
Builder.setExecutable, it will now use the
callable path, not just the execpath.

Fixes https://github.com/bazelbuild/bazel/issues/13189

Unfortunately, call sites that pass PathFragment
are susceptible to the bug, because PathFragment
is inherently ambiguous: if it's a bare file name,
the shell resolves it against PATH, otherwise
against the current directory.

RELNOTES: none

Closes #13192.

PiperOrigin-RevId: 362889202