Bazel's filegroup should participate in toolchain resolution.
This reverts dce861d7c1a13164851fc9070756b946969c5e41, which was a
mistake.

It also removes a test which is now useless. There are no more build
rules (which have the target_compatible_with attribute for target
skipping) which also have toolchain resolution disabled, so this cannot
be tested for.

Fixes #12899 and undoes #13194.