Migrate swift module map to its own action (take 2)
This is in preparation for enabling layering check for `objc_library`.

Layering check requires that a module map be generated.  Currently,
`objc_library` generates a module map to be used by `swift_library`.
The two module maps are incompatible, so we migrate the swift one to
its own action, freein

Eventually, the module map needed by `swift_library` will be migrated
to `swift_library`.

I also had to switch `objc_import` so that it "compiles" its header
and generates a module map.  An external visible effect, however, is
that `objc_import` will now have a `deps` attribute.

PiperOrigin-RevId: 363185731