Handle explicit --desugared_lib_config option passed from unit tests
Some of the unit tests invoke the D8 desugar script with an explicit
option for the configuration. Handle an explicit configuration option
in the D8 desugar scripts by not passing the default configuration in
that case.

Don't allow multiple --desugared_lib_config options to the D8 desugar
invocation Java binary.

RELNOTES: None
PiperOrigin-RevId: 363194875