Remove unnecessary BUILD file in builtins root
Originally, a top-level BUILD file was thought to be needed to help locate the .bzl files, but the current BzlLoadFunction logic doesn't require this. Also, since we abandoned the idea of storing the production builtins files in the install base, we don't need it as an explanatory marker file for any users who go snooping around in there.

This removes the BUILD.builtins file, which used to be renamed to BUILD in the builtins zip. It also replaces a symlink to it with a copy. Finally, the dummy builtins setup in CRCP.Builder is changed to not create it.

Work toward #11437.

PiperOrigin-RevId: 363203944