Improve error messages where there's an IOException reading a source file: report the final status to the user as coming from an IOException, not a "missing file." We still had the right failure detail, but the error message is misleading.
Also print out top-level "missing input" failure messages in keep-going mode, not just no-keep-going.

Noticed as part of this that there were no non-test callers of BuildFailedException#getRootCauses(), and got rid of it. Unfortunately, had to do that in this CL, since without the top-level printing in keep-going mode, we actually weren't ever printing those messages in some cases.

PiperOrigin-RevId: 363204012