Allow builtins also use TransitiveInfoProviderApi in exports parameter of java_common.compile.
Starlark calls java_common.compile and JavaInfo did not populate labels in JavaTransitiveExports provider.

This allows the builtins to use java_common.compile call so that it also populates the labels.

The the JavaTransitiveExportsProvider is going to be removed. Call sites should be cleaned up to use their own aspect to collect the labels.

This CL is a slight regression, so that the cleanup doesn't block Starlarkification of the Java ruleset.

PiperOrigin-RevId: 363360689